services/production-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
psycopg[binary]>=3.2.4
pydantic>=2.0.0
slowapi==0.1.9
httpx>=0.25.0
pika>=1.3.0
python-json-logger==2.0.7

